Add files via upload
Description: Identified the Key factors that leads to Employee Attrition. Built a robust Machine Learning model 
to predict the attrition rate for each employee, to find out who’s more likely to leave the organization. Provided 
Business Recommendations that would help to prevent attrition and to plan in advance the hiring of new 
candidates. 
 Key skills: Python, Pre-processing, EDA, Logistic Regression, Decision Tree, Random Forest Techniques